memoize more stuff
